requirements.txt
google-generativeai
mysql-connector-python
pyyaml
tabulate
